% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{KL}
\alias{KL}
\title{Compute KL divergence between p and q.}
\usage{
KL(p, q)
}
\arguments{
\item{p:}{Actual probability of P given some condition.}

\item{q:}{Initial probability of P.}
}
\description{
KL divergence can be thought of as as expected excess surprise. You think the
chance of U is q, but it's actually p. How surprised are you?
}
\note{
KL divergence is not commutative.
}
