% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{maxVOD}
\alias{maxVOD}
\title{Compute percent of max VOD. When expected disagreement is zero, VOD is
the initial disagreement.}
\usage{
maxVOD(pu_a, pu_b, fun = symmetricKL)
}
\arguments{
\item{fun:}{Divergence measure, defaults to symmetric KL}

\item{pu_a:}{P(U) for model A}

\item{pu_b:}{P(U) for model B}
}
\description{
Compute percent of max VOD. When expected disagreement is zero, VOD is
the initial disagreement.
}
