% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{VoD_log_alt}
\alias{VoD_log_alt}
\title{VoD using KL divergence (log) and geometric mean for P(c) except instead of
taking GMOD of their P(c)s we compute VOI using one and then the other and
average (Tegan's idea)}
\usage{
VoD_log_alt(pu_a, pu_b, puc_a, puc_b, pc_a, pc_b, punotc_a, punotc_b)
}
\description{
VoD using KL divergence (log) and geometric mean for P(c) except instead of
taking GMOD of their P(c)s we compute VOI using one and then the other and
average (Tegan's idea)
}
