% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{VoD_log_gmod}
\alias{VoD_log_gmod}
\title{VoD using KL divergence (log) and geometric mean for P(c)}
\usage{
VoD_log_gmod(pu_a, pu_b, puc_a, puc_b, pc_a, pc_b, punotc_a, punotc_b)
}
\description{
Take the initial difference in terms of symmetric KL divergence, and the
expected disagreement (KL between their P(U|c)s and their P(U|¬c)s, in
expectation), and subtract the latter from the former. Previously we took
the symmetric KL (SKL) between the initial SKL and the expected SKL.
}
\note{
Questions about this? Ask Molly or Ben Tereick.

This is currently the preferred VoD formulation.
}
