% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{VoI_log_updatable}
\alias{VoI_log_updatable}
\title{Compute expected KL divergence, but handle incoherence gracefully.}
\usage{
VoI_log_updatable(pu, puc, pc, punotc)
}
\arguments{
\item{pu:}{P(U)}

\item{puc:}{P(U|c)}

\item{pc:}{P(c)}

\item{punotc:}{P(U|¬c)}
}
\description{
Slightly different framing of VoI_log. For coherent world models, this
yields the same result as VoI_log, but if someone gives an incoherent
response (e.g. if both P(U|c) and P(U|¬c) are greater than P(U)), this will
treat that as information, i.e. thinking about the question made them
update on P(U). Use this if you've collected ALL FOUR VALUES from
respondents. Adapted from VoI_log by Rajashree Agrawal.
}
\note{
Adapted from VoI_log by Rajashree Agrawal.
}
